requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
langchain-community>=0.0.10  # required for Qdrant vector store wrapper
langgraph>=0.0.20
qdrant-client>=1.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
flask>=2.0.0  # required for REST API server
requests>=2.31.0  # required for HTTP requests in interactive mode